+++ app/(tabs)/food.tsx
@@ -1,5 +1,4 @@
 import { StyleSheet } from 'react-native';
-
 import { Text, View } from '@/components/Themed';
 
 export default function TabTwoScreen() {
